requirements.txt
# Autism Storyboard Evaluation Framework - Requirements
# Core dependencies for all evaluation modules

# Deep Learning and ML
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Optional but recommended
pandas>=2.0.0  # For data analysis
jupyter>=1.0.0  # For interactive development